requirements.txt
#sebi_scraper
requests>=2.28.1
beautifulsoup4>=4.11.1
pymongo>=4.3.3
PyPDF2>=3.0.0
spacy>=3.7.5
transformers>=4.24.0
tqdm>=4.64.1
python-dotenv>=0.21.0
torch>=1.13.0
thinc>=8.2.5
python-dotenv
# spaCy small English model
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl


#IOSCO_part2
requests>=2.28.1
beautifulsoup4>=4.11.1
pymongo>=4.3.3
selenium
webdriver-manager
python-dotenv
